requirements.txt
streamlit
opencv-python
faiss-cpu
transformers
torch
sentence-transformers
google-generativeai
python-dotenv